If forced page eviction fails, just give up and hope it works next time.
This allows rec_evict.c to simply set the WT_REF state to WT_REF_MEM after all
failures, and fixes a bug where pages on the forced eviction queue would end up
with state WT_REF_MEM, meaning they could be chosen for eviction multiple
times.

The tradeoff is that when building a tree from scratch that is larger than the
cache size, the eviction thread may never get exclusive access to the single
leaf page in order to evict it.  I've added a yield to the eviction server path
to give application threads a chance to leave the page, but that won't help in
cases of genuine contention (though we've never had a complete answer for that
case).

closes #175