Changes for issue #178.
The underlying problem is btree files weren't completely closed before
they could be re-opened.  This behavior can be triggered by using a
new session on every operation (see the new -S flag to the test/thread
program).  Instead of repeating the dance we do on open, I'm leaving
the btree handles in the connection list for the life of the connection.

The reason I'm going in this direction is that any application that opens
new sessions on each operation (for example, simple scripting language
APIs), will be doing file I/O on every open, and I'd rather avoid that.
(In other words, performance will just suck, instead of being unusable:
it's about 60% slower at the moment).

Most applications don't operate on enough files to care that files aren't
actually discarded, but if you have a long-lived application that opens
thousands of unique files, this approach will eventually see problems
(long file open times as the list of files gets long, and file-descriptor
consumption).

Detailed changes:

When connections are closed, create a session and discard the btree
handles.  I think this fixes a long-standing bug in closing a connection,
that, for any reason, still has open btree handles, we'd drop core when we
tried to evict the pages because we were closing them using the default
session handle.  I don't see any reason we can't allocate a new session
during connection->close, but this is worth careful review.

Change the session close of a btree to be a reference decrement, and
nothing else.  (We don't need the reference right now, but any future
house-cleaning solution will likely need a reference count, so I'm
leaving it in.)

Change __wt_session_lock_btree() to always re-open the tree, even if
it's the first time the session is opening the file.  Now that we're
caching handles in the connection layer, there's no reason to believe
the existing handle is open in any particular configuration.

Move set/clear of the WT_BTREE_OPEN flag into the connection btree handle
code, it's cleaner.  I actually don't think we need the WT_BTREE_OPEN flag
any longer, but I'm not totally convinced of that, and we may need it when
we do cleanup.