Changes for issue #178.
Close/discard all of the files but the schema file first, because writing
those files may result in updates to the schema file.

Updating the schema file may add the schema table to the session's list
of btrees: we're about to discard the underlying schema table file object,
so manually clear that list before closing the last file and discarding
the session handle.