Fixes for bugs found by running test/format multithreaded.
There were two basic problems:

(1) When iterating backwards through a skiplist, we could race with an insert.
In that case, we need to start again from the beginning to update the previous
stack properly.

(2) If eviction fails for a page, we have to assume that eviction has unlocked
the reference: otherwise another thread could have already scheduled it for
eviction again and put it into state WT_REF_EVICTING.