Fix wrong assert macro -- bandaid version
The code in matcher.cpp detects when it is using the wrong assert()
macro in release builds and that code was triggering, preventing
mongod from passing unit tests and making BuildBot useless.  This
fix fixes matcher.cpp only and is just to get the builds going again.
Now looking for which header file did this to us.  More to come ...