Snapshot changes, ref #167:
Add a read-only flag for snapshot load: if we're loading a snapshot for
writing, truncate the file to the snapshot's EOF value, if we're not
loading a snapshot for writing, don't bother reading the avail extent
list.

Add a verify flag rather than use the fragbits array as the verification
boolean.   Add a snapshot-load verification routine that reads the avail
and delete extent lists and adds them to the fragment list, we should not
see those blocks in the tree.