Changes to get ready for maintaining multiple snapshots in the file. We're still only maintaining one snapshot at a time, but most of the code for multiple snapshots is in place.  The actual behavior when deleting a snapshot that's not rolling into the live tree is not yet tested.
Create the WT_SNAPSHOT structure: this structure defines a snapshot's
information, and an array of these structures is shared between the
session, btree and block-manager layers (in the case of the btree and
the block-manager layers, it's referenced by the WT_BTREE->snap field,
because it's serially modified by different threads of control).

Write the code to get/set the array of snapshot entries, rewriting big
chunks of session/session_root.c.

Rename the file-meta "root" string to be "snapshot".

Add a "snapshot" argument to SESSION.open_cursor and SESSION.sync
(support not implemented yet for either one).

Ref #167.