Fix a bug where the full set of modified pages weren't written if the root split as part of a sync (rather than in root page eviction).
Move the eviction code that loops through reconciled pages during
root-split evictions until it finds a simple replacement or emptied page
into the reconciliation code.  Add code to discard the pages created
during a root split, when a subsequent write happens.

Change the page-discard function to know how to free chains of pages,
that is, a single call frees a page and any "split" pages it references.
This trivially simplifies eviction, but the new reconciliation root
split code needs it for real.  Change the eviction code that explicitly
discards split pages to depend on the page-discard function for that.

Change the page-discard function to take a reference to a page and NULL
it, I don't want any risk of code using pages after they've been free'd.