SERVER-5399: Essentially aliasing quit, exit, and function call variations.
Previously, `exit` was being implemented as a code in the DB shell,
while `quit` was being implemented as a shell function.  This meant that
both were doing different things, and were called differently--i.e.,
`exit` could be run as is, whereas `quit()` had to be called as a
function.

Conflicts:

	src/mongo/shell/dbshell.cpp