Fix a bug in the page-reconciliation object tracking code: if blocks are freed after page reconciliation completes, they may be allocated again (for example, as a new overflow item), and appear on the same page, in the same tracked list as the original object.  The tracking software was framed around overflow vs. block objects, and that was wrong: the key issue is if an object is derived from the original page or not, because we only search page-derived objects when reconciling pages and checking if the object has already been entered on the list.  All subsequent allocations must be ignored for the purpose of whether or not a page- related object needs to be freed.  The fix was to completely re-write rec_track.c.
Fix a bug in variable-length column-store overflow items: if an overflow
value has an RLE count greater than 1, and we discarded its backing disk
blocks while the in-memory page still referenced them, and then
re-allocated those blocks to another purpose, cursor walks in the page
might fail because the backing blocks were modified.  (The example was
where the search for a matching overflow value in the list of tracked
objects found a different overflow item, that is, there was more than
one potentially matching overflow item in the list and we re-used the
wrong one, and never re-used the right one).  The fix was to explicitly
re-use the original backing overflow disk blocks instead of trusting in
the tracking object search to find the right entry.

Minor change to enhance the on-disk format debugging display for
variable-length column-store leaf pages.