Remove the timeout in the main select call in mongod
The timeout is only being used to coarsely track elapsed time, but it
causes the process to wake up unnecessarily about 100 times a second,
only for it to discover it has nothing to do and go back to sleep.

Instead, allow the process to sleep until there's work available, and
then query the system clock to figure out how much time has passed.