SERVER-5772 wait specifically for our query, not just any currentOp
The test was failing because the wait for currentOp.inprog to become
non-empty was too generous, and would accept "whatsmyuri" (for example)
instead of the query we are trying to wait for.  Check for our namespace
and for a "$where" query so we only proceed into the real test after the
parallel query has started.