First set of changes for moving all checkpoint/snapshot calls into the session.checkpoint method, ref #235.
Add the "target" configuration to checkpoint, specifies a list of
targets to be checkpointed; checkpointing a list of targets is the
same as checkpointing a database, that is, the checkpoint is a set
of consistent snapshots.

Re-write the checkpoint mehtod documentation to list the new target
config string.

Make __wt_schema_worker signature match __wt_conn_btree_apply, and
__wt_meta_btree_apply (the cfg argument follows the func argument).

Remove cursor declaration in __wt_txn_checkpoint() it's never used.

Add WT_ERR_NOTFOUND_OK macro, allows and corrects WT_NOTFOUND return.