Added fixtures to new unit test framework.
Here's how this take on fixtures with exception handling works:
(Basically, setUp/tearDown mimic the expectations we have for constructors/
destructors in our coding guidelines.)

+ If the fixture throws on setUp() (for instance, because of an ASSERT), then
that test case would be considered failed and tearDown() would not run. The
remainder of the test suite should run fine, assuming calling setUp again in
each test case would build the fixture properly.

+ If the test body code throws because of an ASSERT, then tearDown() will be
issued and the test case would be considered failed. The remainder of the suite
should run fine.

+ If the test body code throws an unexpected assertion, then tearDown() will NOT
run and the entire suite is compromised.

+ tearDown() should be treated as a destructor and therefore should not throw.