If there is an early simple $project, apply it before converting to Documents SERVER-4644
This is a partial fix for SERVER-4644 in that it only works with an
explicit $project and only if that project is supported by the existing
Projection class used to implement the second argument to find().

This also provides a workaround for objects with types that aren't
supported by the Value class (SERVER-4608, SERVER-5718, and SERVER-4968.
Previous behavior was to assert with no possible workaround.

This commit will need some doc updates, in particular in the "Optimizing
Performance" section.