Add support for read-committed isolation.
Operations have transactionally-consistent reads, wherre the snapshot is
updated automatically (when the session reads its first page).

Add a session-level "isolation" setting replacing the one in open_cursor that
was never implemented.  This setting applies to both non-transactional
operations and is the default isolation level for transactions in the session
if non "isolation" setting is specified to WT_SESSION::begin_transaction.

closes #289