Next set of changes for fast-truncate: the ability to delete leaf pages (that don't have overflow items), without reading them.  Ref #247.
Add a new "hidden" type, WT_CELL_ADDR_LNO (leaf no overflow) that marks
a leaf page without overflow items; when we unpack a cell, we set a flag
in the cell structure in this case, the standard unpack type is still
WT_CELL_ADDR.

Add a WT_ADDR.leaf_no_overflow field that flags if the addr references
a leaf page w/o overflow items; reconcilition uses this field to decide
which cell type to use on an internal page.

Add a WT_RECONCILE.ovfl_items flag, and set it when reconciliation finds
an overflow key/value item.  Then, when creating the WT_ADDR for the page,
set the leaf_no_overflow  field if we've found an overflow item.

Update the debugging, verify and salvage routines to match.