Move shared compression tests into a separate class.
The tests in test_compress01 share implementations but have different
subclasses for each compression method.  The structure means that Python
sometimes runs the test on the base class without any compression, which is not
what we want.  Split out the test into a separate class that doesn't inherit
from unittest.TestClass to avoid this.