Another run at fixed-length column-store recno tracking.
The BTREE cursor can be set to an implicit record number, which means
it may not match the the cursor's WT_INSERT reference, and so there may
not be equality in the test for movement through the append list, check
less-than/greater-than and equal for the next//prev movements through
the append list.

Make the prev/next code for setting the record number look the same, add
big comments as to why it works the way it does.

Set the underlying cursor every time we return an implicit record: if a
cursor is positioned to an implicit record, followed by a delete we'll
mess up unless both the interface and lower-level cursors are set to the
returned record.