Detect if truncate cursors are incorrectly ordered, that is, if the start cursor is logically past the stop cursor, add tests to the static test suite for this condition.
Add a private method, WT_CURSOR.compare, only supported for tables and
files that returns a full -1/0/1 comparison of the cursors.

Change the underlying Btree truncation support to instantiate the
start/stop cursors for row-store.  Previously this was done in the
cursor layer, but that's not the layer that depends on it (at least in
the file object case), so do it in the lower layer.

Change WT_CURSOR_NEEDKEY/WT_CURSOR_NEEDVALUE to not return or jump to
the err label, meaning all calls have to be wrapped in WT_ERR/WT_RET
macros.

Change WT_CURSOR.equals methods to consider different objects an error,
rather than simply returning they aren't equal.  Applications passing
in cursors on different objects are seriously misguided, make it an
application error.

Detect if cursor keys aren't set when calling session.truncate, add
tests to the static test suite for this condition.

Add new macro, WT_CURSOR_PRIMARY that returns the primary cursor for a
table.