Another set of changes for fast-truncate (issue #247).
Deal with the possibility that an address-deleted cell must be written
into the file while there are still transactions in the system that may
want a previous version of the page.

Write address-deleted cells based on the commit state and existence of
older snapshots in the system.  If the deletion is uncommitted, write
the original cell information; if the delete is committed, but there are
older snapshots in the system, write the cell's address with a new type
(WT_CELL_ADDR_DEL); if the delete is committed and there are no older
snapshots in the system, write nothing and schedule the free of the
underlying leaf page.   Use the WT_REF.addr field to show the underlying
leaf page has been discarded.

The cache-read function has to handle a new state, where the
address-deleted cell is about to be read from disk, but the underlying
leaf page has been discarded.  Add a function to create an empty leaf
page, shared by the function that instantiates an empty tree and the
cache-read function, for this purpose.

The row-store internal page in-memory instantiation function has to
handle a new cell type: when it finds an address-deleted cell, it has
to re-create the WT_REF state for a deleted node.