SERVER-6816 add a 5 sec time limit to replication batches
Even with a byte cap on replication batches, it is still possible to
batch up a large number of writes which can take a long time to apply.
This commit ensures that batches do not get too large.  5 seconds' worth
of replicated ops over the network will take much less than 5 seconds to
write to the database.