This function solves a problem in reconciliation. The scenario is: 	-  a leaf page that references an overflow item 	-  the item is updated, creating a WT_UPDATE chain 	-  the update is committed 	-  a checkpoint runs, freeing the backing overflow blocks 	-  a snapshot transaction wants the original version of the item
In summary, we may need the original version of an overflow item for a
snapshot transaction, after it's been deleted from the checkpoint tree.
We don't have any way to delay a free of the underlying blocks until a
particular set of transactions exit, and this isn't a common scenario:
read the original overflow item and stash it at the end of the WT_UPDATE
list with a transaction ID guaranteeing no reader will go past the entry.

Reference #310.