SERVER-6671 add a 'v' version field to each oplog document
The version field will allow us to detect the primary's version.
We need to know which version because only newer primary oplog streams
should prevent a secondary from enforcing unique index constraints
in initial sync or recovering states.  The version field will also be
useful in the future when we want to make schema changes in the oplog.