SERVER-2001 filter queries on mongod using new key extraction path
Queries on mongod must filter out documents that do not currently belong
to the shard.  The belongsToMe function now uses the new key extraction
path to make this determination. An optimization is also added to the
clientcursor class so that it can make this determination using a
covered index if available.