Fix LSM index searches.
The main issue was LSM search_near was not always returning the closest key to
the search key, which calling code expects.  It now tries hard to find the
smallest cursor larger than the search key, and only if no larger record exists
does it return the largest record smaller than the search key.

While in the area, clean up the index search code and make it slightly more
efficient.