SERVER-7358 fix issue with mrShardedOutput.js and pre-splitting
An earlier commit ensured that shardcollection only put initial
chunks on the primary shard by default. This is safest, since
there may be concurrent inserts into the collection, but it's not
the behavior that map-reduce expects. This commit allows the caller
of shardcollection to specify a set of initial shards to place
chunks on.  It's safe to use here because the inserts generated by
map-reduce all come from the same source.