We have a panic function: make a failure to clear a hazard pointer return an error and panic the system (using an unpinned page implies corruption).
This means the functions __cursor_func_init, __cursor_func_resolve,
__cursor_leave, __wt_hazard_clear, __wt_page_release and
__wt_stack_release all return error codes, push them all the way up the
call stack.

While I'm there:
	-- don't dump the hazard pointer table if we can't clear a hazard
pointer, it's not useful, by definition our problem is a page that is
NOT in the table.
	-- don't check for a NULL page when clearing a hazard pointer,
it shouldn't ever happen (and we're going to dump core in the field even
if there's an assert there).
	-- simplify the loop in __wt_hazard_close, but no change to the
semantics, the compiler probably did this anyway, but it's easier to
read, regardless.
	-- add a check of the total hazard count to __wt_hazard_close,
it's a good diagnostic check, and we want to know if it's not correct.
	-- don't call __wt_hazard_clear from __wt_hazard_close to clear
hazard references, it provides no added value and is just going to make
things messier in the diagnostic output.