Another set of changes to make checkpoints outside of the eviction server thread work.  The last set of changes had (at least) two problems:
First, the test to avoid selecting a dirty page for eviction was
insufficient.  It appeared in __evict_walk_file, which is before
eviction has exclusive access, the test should still appear there to
avoid selecting pages there is no hope of evicting, but should also
appear in __rec_review after the page is locked down to ensure no pages
modified after selection, but before final review, are written.

Second, the test to avoid selecting a dirty page for eviction wasted
performance because we couldn't select a page ever considered for
modification (regardless of whether or not it was actually modified),
because racing with the checkpoint thread reviewing an internal page
that referenced the evicted page could drop core when the evicted page's
modification structure disappeared.

This set of changes adds a test in __rec_review to resolve the first
problem.

The second problem is trickier, and messes with page states (oh joy!).
The key is to inform page reconciliation if it's being called by the
eviction server thread or a checkpoint thread.  In the case of being
called by the eviction server thread, any child page of an internal page
that's in the WT_REF_LOCKED state is in a stable, in-memory state,
because the calling thread set the WT_REF_LOCKED state.  In the case of
being called by the checkpoint thread, any child page of an internal
page that's in the WT_REF_LOCKED state is in a temporary state because
it's been locked by eviction.  It will either be evicted (and the page
state reset to WT_REF_DISK), or skipped (and the page state reset to
WT_REF_MEM).  Regardless, the reconciliation within the checkpoint
thread has to wait on that state change.

This is all built on top of a set of flags passed into reconciliation,
which additionally offers better control over skipping updates on a
page.  We now explicitly flag if reconciliation should (1) quit early
if unable to entirely clean a page, set by the eviction code, (2) panic
if it's unable to entirely clean a page, set by sync during file close
and by salvage, or (3) not worry about it, set by the checkpoint passes.

Reference #419.