Simplify WT_PANIC_XXX and friends.
I needed a return version, which made 3, but we only have 2 cases: (1)
one that returns WT_PANIC without any error value for the associated
message, and (2) one that doesn't return WT_PANIC (ever), but does have
an error value for the associated message.

Case #1 is now WT_PANIC_ERR, case #2 is WT_PANIC_RETX, and that leaves
WT_PANIC_ERRX for a version that doesn't return WT_PANIC (ever) but does
have an error value for the associated message, and WT_PANIC_RET for a
version that returns WT_PANIC and has an error value for the associated
message, if we ever need them.