Instead of skipping eviction of empty pages, set the parent page's WT_REF to WT_REF_DELETED, that way we can discard the page.  Reference #437.
If we instantiate an empty page (and it's really empty, for example, it
was instantiated as part of a search), and we eventually discard it
without modifying it, we have to reset its WT_REF status to empty, not
to on-disk, because it has no disk address.

Don't cache a reference to the page's modify structure, the call into
reconciliation can swap out the page's modify structure for a new one.