In order to mmap read-only checkpoints, we need a place to store a per-checkpoint mapping.  We don't currently have a per-checkpoint structure in the block manager, we only have the per-file WT_BLOCK. Add the WT_BM structure, which is a per-checkpoint structure.
While we're in the area...

Share the underlying WT_BLOCK structure between WT_BM structures, this
means checkpoints into the file will share a single file descriptor.

Give the WT_BM structure a set of method pointers, this is the change
that we'd need to plug in other block managers.

Some minor cleanups of size_t/uint32_t values.