Include the default file configuration when opening a file.
The defaults are included because underlying objects have persistent
configuration information stored in the metadata file.  If defaults are
included in the configuration, we can add new configuration strings
without upgrading the metadata file or writing special code in case a
configuration string isn't initialized, as long as the new configuration
string has an appropriate default value.