Attempt to improve performance of primary when secondaries are pulling oplog data from the primary.
On a low latency network, if secondaries are "caught up" they may pull very small
getmore batches from the primary.  This will result in a medium high getmore ops per
second rate on the primary and some lock contention for local db lock. In addition
it is better for the secondary to apply batches that aren't tiny (given its parallelism
model).

There was already a notion of this with AwaitData in the primary-side code however
very often on the follow-up getmore request there is at least one or two new ops to
pick up so the short sleep there doesn't really kick in so much.

The change is to sleep for a short interval before asking for more oplog if we are
caught up.  We infer we are "caught" up if the last batch received was not very big.