Implementation of a fast-path in class Matcher for oplog queries.
Those queries are of the form:

  { ts : { $gte : Timestamp(<val>) } }

Profiler data (windows, before->after):

processGetMore
 (Exclusive samples - getmore is much faster than the 1 secondary that was running...)
 2.28% -> 1.49%

Cursor::currentMatches fraction of processGetMoreTime
 (Inclusive samples)
 50%->36%

So empirically a big difference.

I believe Matcher is being refactored in general. So perhaps this can go away when
that is done but oplog receivedGetMore performance should be measured against this
implementation at that time before considering removing this.