Fixes for the memory accounting of split-merge pages.
The eviction server was using the cache_inuse routines on exit, which
don't complain if there are more bytes/pages evicted than are in memory,
and that's not correct, change it to do an exact comparison.  This was
hiding some problems in our page count.  Fix the page count problems by:
	... adding an "allocate a new WT_PAGE" function that increments
the count of pages in the system and calling it from everywhere (we
weren't counting split-merge pages created during reconciliation, or the
pages we created when opening empty trees);
	... change the split-merge collapse code to call __wt_page_out
to discard pages instead of freeing them explicitly;
	... rename WT_CACHE.pages_read to be pages_inmem, matching the
bytes_inmem name and matching its new function of tracking all pages,
no matter their origin.

We were tracking cache "bytes in memory" in two variables: "bytes read"
was tracked in WT_CACHE.bytes_read and "bytes subsequently added in
memory" in WT_CACHE.bytes_inmem, and we added them together to get a
final answer.  Change WT_CACHE.bytes_inmem to be the bytes in memory,
regardless of whether it was read in, a created page, or subsequently
added memory, by removing WT_CACHE.bytes_read and changing the
__wt_page_inmem function to call __wt_cache_page_inmem_incr instead of
having a separate __wt_cache_page_read function.

Change __wt_page_modify_init to increment the page's footprint when the
WT_PAGE_MODIFY structure is first allocated.

Add a reconciliation function to create split-merge pages: abstract out
the common code and set the page's memory footprint.