Add a new WT_EXTENSION_API method that returns the default session for a WT_CONNECTION, we need a session to use the other WT_EXTENSION_API methods.
Remove the "exclusive" argument to the WT_DATA_SOURCE.create method:
we handle it internally and it's not documented anywhere (besides it's
not standard for the WT_SESSION.create method, if the WT_SESSION.create
method needs it, it should retrieve it from the configuration information).

Change the WT_DATA_SOURCE's schema create function to use the
session.create configuration, instead of the file_meta configuration
the file: type uses; this is still wrong, but at least we're passing
the correct application-specified configuration information to the
extension functions now.