SERVER-2114 Don't use select() timeouts for fast coarse timing
On Linux, the CLOCK_MONOTONIC_COARSE clock is very fast to read, as
little as 25ns on my little laptop. It's resolution is limited to the
kernel tick resolution, which on most servers will be 100hz; the current
kludge uses select() timeouts of 10ms, so this should be acceptable.

On other platforms, the penalty for bona fide syscalls should be nearly
negligable - the real syscalls on Linux take around 1 microsecond on my
little laptop, and I can't imagine they would be that much worse on
Windows/OSX/Solaris. Even if it is worse, avoiding 100 wakeups a second
from select() may very well offset the penalty.