Give the LSM tree its own data-source statistics structure back, move the bloom filter statistics back into that structure.
Change the underlying stat macros to check if statistics are configured
at the lowest (WT_STAT_XXX) levels: this means we potentially check to
see if statistics are configured a lot more (every time a stat field is
updated), but it ensures we don't actually maintain the statistics
unless they're configured.  Trust the compiler will optimize out the
extra tests for configuration.